+++ backend/app/models/channel_supplier.py
@@ -1,5 +1,5 @@
 """渠道供应商模型"""
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey, Enum as SQLEnum
+from sqlalchemy import Column, Integer, BigInteger, String, Boolean, DateTime, ForeignKey, Enum as SQLEnum
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.core.database import Base
@@ -18,8 +18,8 @@ class ChannelSupplier(Base):
     """渠道供应商表"""
     __tablename__ = "channel_suppliers"
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    tenant_id = Column(Integer, ForeignKey("tenants.id"), nullable=False, index=True, comment="所属甲方ID")
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    tenant_id = Column(BigInteger, ForeignKey("tenants.id"), nullable=False, index=True, comment="所属甲方ID")
     channel_type = Column(SQLEnum(ChannelTypeEnum), nullable=False, index=True, comment="渠道类型：sms/rcs/whatsapp/call")
     supplier_name = Column(String(200), nullable=False, comment="供应商名字")
     api_url = Column(String(500), nullable=False, comment="供应商接口地址")
